requirements.txt
acme==1.12.0
aiohttp==3.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
asyncio-dgram==2.1.2
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
CacheControl==0.12.11
cachelib==0.1.1
cairocffi==1.4.0
CairoSVG==2.5.2
certbot==1.12.0
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
cleo==2.0.1
click==8.1.3
ConfigArgParse==1.2.3
configobj==5.0.6
crashtest==0.4.1
cryptography==38.0.4
cssselect2==0.7.0
dbus-python==1.2.16
defusedxml==0.7.1
discord==2.1.0
discord.py==2.1.0
distlib==0.3.6
distro==1.5.0
dnspython==2.2.1
dominate==2.7.0
dulwich==0.20.50
eventlet==0.33.1
exceptiongroup==1.0.4
filelock==3.8.2
Flask==2.2.2
Flask-API==3.0.post1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
frozenlist==1.3.3
greenlet==1.1.3.post0
gyp==0.1
h11==0.14.0
html5lib==1.1
httplib2==0.18.1
idna==3.4
importlib-metadata==4.13.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.2
josepy==1.2.0
jsonschema==4.17.3
keyring==23.11.0
kombu==5.2.4
lockfile==0.12.2
lxml==4.9.1
MarkupSafe==2.1.1
mcrcon==0.7.0
mcstatus==10.0.1
mock==2.0.0
more-itertools==9.0.0
msgpack==1.0.4
multidict==6.0.2
numpy==1.23.5
oauthlib==2.0.6
opengsq==1.4.3
outcome==1.2.0
packaging==22.0
parsedatetime==2.6
pbr==5.11.0
pexpect==4.8.0
pika==1.3.1
Pillow==9.3.0
pkginfo==1.9.2
platformdirs==2.6.0
poetry==1.3.1
poetry-core==1.4.0
poetry-plugin-export==1.2.0
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pycparser==2.21
pycurl==7.43.0.6
PyGObject==3.38.0
PyMySQL==1.0.2
pyOpenSSL==22.1.0
pyRFC3339==1.1
pyrsistent==0.19.2
PySimpleSOAP==1.16.2
PySocks==1.7.1
python-apt==2.2.1
python-debian==0.1.39
python-debianbts==3.1.0
python-magic==0.4.20
pytz==2021.1
rapidfuzz==2.13.6
reportbug==7.10.3+deb11u1
requests==2.28.1
requests-oauthlib==0.8.0
requests-toolbelt==0.9.1
SecretStorage==3.3.3
selenium==4.6.1
shellingham==1.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
stem==1.8.1
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.6
trio==0.22.0
trio-websocket==0.9.2
trove-classifiers==2022.12.1
ufw==0.36
urllib3==1.26.13
vine==5.0.0
virtualenv==20.17.1
visitor==0.1.3
webencodings==0.5.1
Werkzeug==2.2.2
wsproto==1.2.0
WTForms==3.0.1
xvfbwrapper==0.2.9
yarl==1.8.1
zipp==3.11.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.0.1
zope.interface==5.2.0
